Restructured Flow
Previously, the program had a lot of needless lines of code due to the
flow. We were essentially writing the same (but surprisingly different?)
drawing, calculations, and publishing to the robot code twice. I changed
around the flow inside of findTower() to find the best rectangle (by
choosing the goal with the largest width) and then to draw. This should
hopefully make the code easier to read and understand.

We also now draw a cyan box around other goals and a orange box (with a
red center) above our closest goal. I'll wait to test this before
merging back into main. I had some issues with bracket misalignment that
makes me afraid to commit to master.