Probability of scoring for X
An example of how to calculate the probability of scoring. The data
needed is the variance in the X and Y of the ball hit when the robot
fires from the same distance, same angle, and same power. I.e. the only
thing effecting the shot is the actual variance in the hardware itself.